requirements.txt
lean
pytest
flake8
finnhub-python
pandas
pyarrow
alpaca-trade-api
python-dotenv